+++ app/src/main/java/woowacourse/shopping/data/dataSource/local/recent/RecentProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.model
+package woowacourse.shopping.data.dataSource.local.recent
 
 data class RecentProductEntity(
     val productId: Long,
