+++ app/src/main/java/woowacourse/shopping/data/repository/local/CartRepositoryImpl.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.repository.local
 import com.example.domain.model.CartProduct
 import com.example.domain.model.Product
 import com.example.domain.repository.CartRepository
-import woowacourse.shopping.data.sql.cart.CartDao
+import woowacourse.shopping.data.dataSource.local.cart.CartDao
 
 class CartRepositoryImpl(
     private val cartDao: CartDao
