+++ app/src/main/java/woowacourse/shopping/feature/detail/dialog/CounterDialog.kt
@@ -9,8 +9,8 @@ import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
 import woowacourse.shopping.R
 import woowacourse.shopping.common_ui.CounterView
+import woowacourse.shopping.data.dataSource.local.cart.CartDao
 import woowacourse.shopping.data.repository.local.CartRepositoryImpl
-import woowacourse.shopping.data.sql.cart.CartDao
 import woowacourse.shopping.databinding.DialogCounterBinding
 import woowacourse.shopping.model.ProductUiModel
 import woowacourse.shopping.util.getParcelableCompat
