+++ app/src/test/java/woowacourse/shopping/data/MockRemoteProductRepositoryTest.kt
@@ -6,9 +6,9 @@ import com.example.domain.model.Product
 import com.example.domain.repository.ProductRepository
 import org.junit.Before
 import org.junit.Test
+import woowacourse.shopping.data.dataSource.remote.MockProductRemoteService
 import woowacourse.shopping.data.repository.remote.MockRemoteProductRepositoryImpl
-import woowacourse.shopping.data.service.MockProductRemoteService
-import java.lang.Thread.sleep
+import java.util.concurrent.CountDownLatch
 
 internal class MockRemoteProductRepositoryTest {
     private lateinit var mockRemoteProductRepository: ProductRepository
@@ -26,15 +26,17 @@ internal class MockRemoteProductRepositoryTest {
     fun `처음 20개의 상품을 불러온다`() {
         // given
         var actual: List<Product> = listOf()
+        val latch = CountDownLatch(1)
 
         // when
         mockRemoteProductRepository.fetchFirstProducts(
             onSuccess = {
                 actual = it
+                latch.countDown()
             },
             onFailure = {}
         )
-        sleep(1000)
+        latch.await()
 
         // then
         val expected = productsDatasource.subList(0, 20)
@@ -42,19 +44,21 @@ internal class MockRemoteProductRepositoryTest {
     }
 
     @Test
-    fun `20번 아이디 상품까지 받은 상태에서 추가적으로 더 불러온다`() {
+    fun `20개의 상품을 받은 상태에서 다음 20개의 상품을 추가적으로 더 불러온다`() {
         // given
         var actual: List<Product> = listOf()
+        val latch = CountDownLatch(1)
 
         // when
         mockRemoteProductRepository.fetchNextProducts(
             20L,
             onSuccess = {
                 actual = it
+                latch.countDown()
             },
             onFailure = {}
         )
-        sleep(1000)
+        latch.await()
 
         // then
         val expected = productsDatasource.subList(20, 40)
@@ -65,16 +69,18 @@ internal class MockRemoteProductRepositoryTest {
     fun `40번 아이디 상품까지 받은 상태에서 추가적으로 더 불러온다`() {
         // given
         var actual: List<Product> = listOf()
+        val latch = CountDownLatch(1)
 
         // when
         mockRemoteProductRepository.fetchNextProducts(
             40L,
             onSuccess = {
                 actual = it
+                latch.countDown()
             },
             onFailure = {}
         )
-        sleep(1000)
+        latch.await()
 
         // then
         val expected = productsDatasource.subList(40, 41)
